fix: Api 관련 class 및 interface 이관 및 response 정상 받아오는 상태로 수정
